+++ cirque/nodes/wifiapnode.py
@@ -13,17 +13,15 @@
 # limitations under the License.
 
 import os
-import time
 import random
+import time
 
-from subprocess import PIPE
-
-import cirque.common.utils as utils
 from cirque.capabilities.wificapability import WiFiCapability
 from cirque.common.exceptions import (
     ContainerExecError,
     IpNetnsExecError,
 )
+import cirque.common.utils as utils
 from cirque.nodes.dockernode import DockerNode
 
 RUNTIME_NAMESPACE = "/var/run/netns"
@@ -31,24 +29,89 @@ CHAR_SRC = "ABCDEFGHIJKLMNOPQRSTUVWXYZ123456789"
 
 
 class WiFiAPNode(DockerNode):
+  """Simulated Wi-Fi Access Point node supporting dual-mode Virtual Wi-Fi and legacy hostapd/dnsmasq.
 
-  def __init__(self,
-               docker_client,
-               ssid=None,
-               password=None,
-               container_name=None,
-               base_image="mac80211_ap_image"):
-    super().__init__(docker_client, node_type="wifi_ap", base_image=base_image)
+  In default userspace Virtual Wi-Fi mode (`use_virtual_wifi_tcp=True`):
+    - Registers the Access Point SSID, WPA2 PSK, BSSID (`02:00:00:00:01:01`),
+      and gateway addresses (`10.0.1.1/24`, `fd11:22::1/64`) with the host
+      `VirtualWiFiServer`.
+    - Does not require `mac80211_hwsim.ko` or `mac80211_ap_image` (gracefully
+      falls back to `project-chip/chip-cirque-device-base` if
+      `mac80211_ap_image`
+      is not built locally).
+
+  In legacy kernel mode (`use_virtual_wifi_tcp=False` or
+  `CIRQUE_USE_LEGACY_HWSIM=1`):
+    - Moves a `mac80211_hwsim` `phyX` radio into the container's network
+      namespace and starts `hostapd` and `dnsmasq` inside the container.
+  """
+
+  def __init__(
+      self,
+      docker_client,
+      ssid=None,
+      password=None,
+      **kwargs,
+  ):
+    """Initializes the WiFiAPNode container and registers the AP on VirtualWiFiServer."""
+    container_name = kwargs.get("container_name")
+    base_image = kwargs.get("base_image", "mac80211_ap_image")
+    capabilities = kwargs.get("capabilities")
+    use_virtual_wifi_tcp = kwargs.get("use_virtual_wifi_tcp")
+    if use_virtual_wifi_tcp is None:
+      legacy_env = (
+          os.environ.get("CIRQUE_USE_LEGACY_HWSIM", "0").strip().lower()
+      )
+      use_virtual_wifi_tcp = legacy_env not in ("1", "true", "yes")
+    resolved_image = self._resolve_ap_base_image(docker_client, base_image)
+    init_caps = list(capabilities) if capabilities else []
+    super().__init__(
+        docker_client,
+        node_type="wifi_ap",
+        capabilities=init_caps,
+        base_image=resolved_image,
+    )
     random.seed(time.time())
     self.ssid = ssid
     self.password = password
     self.container_name = container_name
-    self.wifi_capability = WiFiCapability()
-    self.capabilities.append(self.wifi_capability)
+    self.wifi_capability = None
+    for cap in self.capabilities:
+      if getattr(cap, "name", "") == "WiFi":
+        self.wifi_capability = cap
+        self.wifi_capability.is_ap = True
+        self.wifi_capability.auto_connect = True
+        break
+    if self.wifi_capability is None:
+      self.wifi_capability = WiFiCapability(
+          use_virtual_wifi_tcp=use_virtual_wifi_tcp,
+          auto_connect=True,
+          is_ap=True,
+      )
+      self.capabilities.append(self.wifi_capability)
     if not self.ssid:
       self.ssid = self.get_wifi_ssid()
     if not self.password:
       self.password = self.get_wifi_password()
+    if (
+        self.wifi_capability.use_virtual_wifi_tcp
+        and WiFiCapability._SHARED_VIRTUAL_SERVER is not None
+    ):
+      WiFiCapability._SHARED_VIRTUAL_SERVER.register_ap(
+          ssid=self.ssid,
+          psk=self.password,
+      )
+
+  @staticmethod
+  def _resolve_ap_base_image(docker_client, requested_image):
+    try:
+      docker_client.images.get(requested_image)
+      return requested_image
+    except Exception:  # pylint: disable=broad-exception-caught
+      fallback = os.environ.get(
+          "CHIP_CIRQUE_BASE_IMAGE", "project-chip/chip-cirque-device-base"
+      )
+      return fallback
 
   def __del__(self):
     if hasattr(self, "container") and self.container:
@@ -60,14 +123,23 @@ class WiFiAPNode(DockerNode):
 
   @property
   def description(self):
-    inspection = self.inspect()
-    return {"type": "wifi_ap", "ssid": self.ssid, "psk": self.password}
+    return {
+        "type": "wifi_ap",
+        "ssid": self.ssid,
+        "psk": self.password,
+        "ipv4_addr": "10.0.1.1",
+        "ipv6_addr": "fd11:22::1",
+    }
 
   def stop(self):
     if hasattr(self, "container") and self.container:
       self.__teardown()
-      self.container.stop()
-      del self.container
+      self.container.stop(timeout=2)
+      try:
+        self.container.remove(force=True)
+      except Exception:  # pylint: disable=broad-exception-caught
+        pass
+      self.container = None
 
   def run(self):
     kwargs = {"stdin_open": True, "cap_add": ["NET_ADMIN"]}
@@ -77,6 +149,19 @@ class WiFiAPNode(DockerNode):
 
     super().run(**kwargs)
     self.logger.info("Creating WiFi AP node: {}".format(self.name))
+    if self.wifi_capability.use_virtual_wifi_tcp:
+      if WiFiCapability._SHARED_VIRTUAL_SERVER is not None:
+        WiFiCapability._SHARED_VIRTUAL_SERVER.register_ap(
+            ssid=self.ssid,
+            psk=self.password,
+            ap_id=self.name or "ap0",
+        )
+      try:
+        self.__setup_network_forward_rules()
+      except Exception:  # pylint: disable=broad-exception-caught
+        pass
+      return
+
     self.__setup_namespace_networking_env()
     self.__setup_network_forward_rules()
     self.__update_hostapd_ssid()
@@ -86,17 +171,25 @@ class WiFiAPNode(DockerNode):
     utils.sleep_time(self.logger, 2, "creating WiFi AP node")
 
   def get_wifi_ssid(self):
-    return self.ssid if self.ssid else \
-        "-".join(["wifiap",
-                  "".join(random.choice(CHAR_SRC) for _ in range(5))])
+    return (
+        self.ssid
+        if self.ssid
+        else "-".join(
+            ["wifiap", "".join(random.choice(CHAR_SRC) for _ in range(5))]
+        )
+    )
 
   def get_wifi_password(self):
-    return self.password if self.password else \
-        "".join(random.choice(CHAR_SRC) for _ in range(8))
+    return (
+        self.password
+        if self.password
+        else "".join(random.choice(CHAR_SRC) for _ in range(8))
+    )
 
   def __run_hostapd(self):
     ret = self.container.exec_run(
-        "/usr/sbin/hostapd -B /etc/hostapd/hostapd.conf")
+        "/usr/sbin/hostapd -B /etc/hostapd/hostapd.conf"
+    )
     if ret.exit_code != 0:
       raise ContainerExecError("unable to start up hostapd server!!")
 
@@ -114,8 +207,9 @@ class WiFiAPNode(DockerNode):
     for command in network_commands:
       ret = utils.netns_run(self.logger, command, self.name)
       if ret.returncode != 0:
-        raise IpNetnsExecError("failed to run cmd: {}\n {}".format(
-            command, ret.stderr))
+        raise IpNetnsExecError(
+            "failed to run cmd: {}\n {}".format(command, ret.stderr)
+        )
 
     ret = self.container.exec_run("route add -net 10.0.1.0/24 gw 10.0.1.1")
     if ret.exit_code != 0:
@@ -124,34 +218,41 @@ class WiFiAPNode(DockerNode):
   def __setup_network_forward_rules(self):
     forward_rules = [
         "iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE",
-        "iptables -A FORWARD -i eth0 -o wlan0 -m \
-                state --state RELATED,ESTABLISHED -j ACCEPT",
+        (
+            "iptables -A FORWARD -i eth0 -o wlan0 -m "
+            "state --state RELATED,ESTABLISHED -j ACCEPT"
+        ),
         "iptables -A FORWARD -i wlan0 -o eth0 -j ACCEPT",
     ]
 
     for rule in forward_rules:
       ret = self.container.exec_run(rule)
-      if ret.exit_code != 0:
+      if ret.exit_code != 0 and not self.wifi_capability.use_virtual_wifi_tcp:
         raise ContainerExecError("unable to set forward rule: {}".format(rule))
 
   def __update_hostapd_ssid(self):
-    ret = self.container.exec_run("sed -i 's/ssid=/ssid={}/' \
-                /etc/hostapd/hostapd.conf".format(self.ssid))
+    ret = self.container.exec_run(
+        "sed -i 's/ssid=/ssid={}/' /etc/hostapd/hostapd.conf".format(self.ssid)
+    )
     if ret.exit_code != 0:
-      raise ContainerExecError("failed to set wifi ap ssid: {}".format(
-          self.ssid))
+      raise ContainerExecError(
+          "failed to set wifi ap ssid: {}".format(self.ssid)
+      )
 
   def __update_hostapd_password(self):
     ret = self.container.exec_run(
-        "sed -i 's/wpa_passphrase=/wpa_passphrase={}/' \
-                /etc/hostapd/hostapd.conf".format(self.password))
+        "sed -i 's/wpa_passphrase=/wpa_passphrase={}/' "
+        "/etc/hostapd/hostapd.conf".format(self.password)
+    )
     if ret.exit_code != 0:
-      raise ContainerExecError("failed to set wifi ap password: {}".format(
-          self.pasword))
+      raise ContainerExecError(
+          "failed to set wifi ap password: {}".format(self.password)
+      )
 
   def __teardown(self):
-    self.logger.info("stopping hostapd...")
-    self.container.exec_run("killall hostapd")
-    self.logger.info("stopping dnsmasq...")
-    self.container.exec_run("killall dnsmasq")
+    if not self.wifi_capability.use_virtual_wifi_tcp:
+      self.logger.info("stopping hostapd...")
+      self.container.exec_run("killall hostapd")
+      self.logger.info("stopping dnsmasq...")
+      self.container.exec_run("killall dnsmasq")
     self.wifi_capability.disable_capability(self)
